pyproject.toml
[project]
name = "receipt-processor"
version = "1.0.0"
description = "AI-powered receipt processing application with Gmail integration and banking automation"
requires-python = ">=3.11"
authors = [
    { name = "Brian Kaplan", email = "kaplan.brian@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
keywords = ["receipt", "processing", "ai", "gmail", "banking", "ocr", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web Framework
    "flask>=2.3.0",
    "gunicorn>=21.2.0",
    
    # Google APIs
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-cloud-vision>=3.10.0",
    "gspread>=6.1.0",
    
    # Database
    "pymongo>=4.13.0",
    
    # Cloud Storage
    "boto3>=1.38.0",
    "botocore>=1.38.0",
    
    # AI and ML
    "openai>=1.88.0",
    "transformers>=4.50.0",
    "sentence-transformers>=4.0.0",
    "torch>=2.0.0",
    "accelerate>=0.30.0",
    
    # OCR and Image Processing
    "pillow>=11.2.0",
    "pytesseract>=0.3.13",
    
    # HTTP and Async
    "requests>=2.32.0",
    "httpx>=0.28.0",
    "aiohttp>=3.12.0",
    "aiofiles>=23.2.0",
    
    # HTML Parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "cryptography>=45.0.0",
    "tqdm>=4.66.0",
    
    # OAuth and Security
    "oauth2client>=4.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/briankaplan/receipt-processor"
Repository = "https://github.com/briankaplan/receipt-processor.git"
Issues = "https://github.com/briankaplan/receipt-processor/issues"
Documentation = "https://github.com/briankaplan/receipt-processor/blob/main/README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py311"]
include = ".pyi?$"
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true